Moved the files to appropriate folders
